Add sleep between request when updating timestamp
Updating timestamp is a hack to resume logs fetching from 1 microsecond after
last log fetched from last iteration. It may miss logs, but chance of that
happening is very log.

The only other correct way to do this is to keep some state which is an
overkill. This simplifies logic. To not keep on updating timestamp, added a
sleep of 1 second between 2 request fetches if timestamp is updated